+++ dune/grid/albertagrid/gridfactory.hh
@@ -15,8 +15,6 @@
 #include <map>
 #include <memory>
 
-#include <dune/common/to_unique_ptr.hh>
-
 #include <dune/geometry/referenceelements.hh>
 
 #include <dune/grid/common/gridfactory.hh>
@@ -290,7 +288,7 @@ namespace Dune
      *  \note ALBERTA's grid factory provides a static method for freeing the
      *        grid (destroyGrid).
      */
-    ToUniquePtr<Grid> createGrid ()
+    std::unique_ptr<Grid> createGrid ()
     {
       macroData_.finalize();
       if( macroData_.elementCount() == 0 )
@@ -300,7 +298,7 @@ namespace Dune
       assert( macroData_.checkNeighbors() );
       macroData_.checkCycles();
       ProjectionFactory projectionFactory( *this );
-      return new Grid( macroData_, projectionFactory );
+      return std::make_unique<Grid>( macroData_, projectionFactory );
     }
 
     /** \brief destroy a grid previously obtain from this factory