+++ dune/grid/uggrid/uggridfactory.hh
@@ -14,7 +14,6 @@
 #include <vector>
 
 #include <dune/common/fvector.hh>
-#include <dune/common/to_unique_ptr.hh>
 
 #include <dune/grid/common/boundarysegment.hh>
 #include <dune/grid/common/gridfactory.hh>
@@ -216,7 +215,7 @@ namespace Dune {
 
        The receiver takes responsibility of the memory allocated for the grid
      */
-    virtual ToUniquePtr<UGGrid<dimworld>> createGrid();
+    virtual std::unique_ptr<UGGrid<dimworld>> createGrid();
 
     static const int dimension = UGGrid<dimworld>::dimension;